Move SF related code outside of api modules
Notification backends should be plugabble which means that
API code should not be changed by adding/removing new notification
drivers
